server/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bson==0.5.10
certifi==2025.1.31
click==8.1.8
distro==1.9.0
dnspython==2.7.0
easyocr==1.7.2
email_validator==2.2.0
fastapi==0.115.8
filelock==3.17.0
fsspec==2025.2.0
groq==0.16.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
Jinja2==3.1.5
lazy_loader==0.4
MarkupSafe==3.0.2
motor==3.7.0
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.3
numpy==2.2.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pillow==11.1.0
pipfile==0.0.2
pyclipper==1.3.0.post6
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
scikit-image==0.25.1
scipy==1.15.1
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
sympy==1.13.3
tifffile==2025.1.10
toml==0.10.2
torch==2.2.2
torchvision==0.17.2
typing_extensions==4.12.2
uvicorn==0.34.0
